Create entry docs update
