+++ lib/stack/contentType/entry/index.js
@@ -277,7 +277,12 @@ export function Entry (http, data) {
    * const client = contentstack.client()
    * const entry  = {
    *  title: 'Sample Entry',
-   *  url: '/sampleEntry'
+   *  url: '/sampleEntry',
+   *  file: 'asset_uid', // for single asset pass asset uid to entry asset field value
+   *  multiple_file = ['asset_uid_1', 'asset_uid_2'], // for multiple asset pass array of asset uid to entry asset field values
+   *  reference: reference.uid, // for single reference pass reference uid to entry reference field value
+   *  multiple_reference: ['reference_uid_1', 'reference_uid_2'], // for multiple reference pass array of reference uid to entry reference field values
+   *  multiple_content_type_reference: [{_content_type_uid: 'content_type_uid_1', uid: 'reference_uid_1'}, {_content_type_uid: 'content_type_uid_2', uid: 'reference_uid_2'}] // for multiple reference pass array of reference uid to entry reference field values
    * }
    * client.stack().contentType('content_type_uid').entry().create({ entry })
    * .then((entry) => console.log(entry))
